requirements.txt
pandas
numpy
matplotlib
seaborn
jupyterlab
python-dotenv
scikit-learn
scipy
phik
catboost
optuna
ipywidgets
fastapi
uvicorn
mlflow
psycopg
psycopg_binary
psycopg2-binary
prometheus_client
prometheus-fastapi-instrumentator
boto3
implicit
python-multipart